+++ src/lib/Style.css
@@ -1,0 +1,41 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    }
+.carrossel {
+    justify-content: center;
+    align-items: center;
+    display: flex;
+    margin: 5px;
+    overflow: hidden;
+    position: relative;
+    height:440px;
+    
+    }
+.bloco {
+    align-items: center;
+    display: flex;
+    scroll-snap-type: x mandatory;
+    width: 100%;
+    height: 355px;
+    overflow: hidden;
+    position: relative;
+    
+    
+}
+
+.bloco div {
+    flex: 0 0 auto;
+    display: flex;
+    width: 150px;
+    height: 300px;
+    margin: 10px;
+    cursor: pointer;
+    border-radius: 5px;
+    transition: transform 0.3s ease-in-out, box-shadow 0.3s ease-in-out;
+}
+.bloco div:hover{
+    transform: scale(1.15); /* Aumenta o tamanho ao passar o mouse */
+    box-shadow: 0 0 10px rgba(0, 0, 0, 1); /* Adiciona uma sombra ao passar o mouse */
+}
