+++ src/lib/carrossel.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">    
+    <title>Componente</title>
+    <link rel = "styleshet" href="Style.css">
+    
+    
+</head>
+<body>
+    <div class="carrossel">
+        <div class="bloco" id="carrossel">
+        </div>
+    </div>
+    
+    <script src = "script.js" ></script>
+    
+</body>
+</html>
